backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "clerk-backend-api>=4.2.0",
    "fastapi[standard]>=0.125.0",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "svix>=1.83.0",
    "uvicorn>=0.38.0",
]
